requirements.txt
# Core
tensorflow==2.15.0              # Stable TF version with Keras integration
tensorflow-text==2.15.0        # Text processing (matches TF version)
tensorflow-datasets==4.9.4     # Compatible with TF 2.15
numpy==1.24.3                  # Avoid TF 2.15+ compatibility issues

# Text preprocessing
nlpaug==1.1.3                  # Compatible with TF 2.x
nltk==3.8.1                    # Tokenization, stopwords
spacy==3.7.2                   # Advanced text processing
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl  # English model for spaCy

# Utilities
pandas==2.0.3                  # Data handling
scikit-learn==1.3.2            # Metrics, data splits
tqdm==4.66.1                   # Progress bars

# Optional for API deployment
fastapi==0.109.0               # REST API (tested with Python 3.9+)
uvicorn==0.27.0                # ASGI server

# Edge deployment
tensorflow-model-optimization==0.7.5  # Model pruning/quantization